+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Sequence IDs are now reported for iLocus and miLocus tables.
 - New script for creating a piLocus summary table.
 - Recipes for chick pea, cabbage, and soybean.
-- A `--shuffled` option to several scripts for reading from `*.shuffled.*iloci.tsv`.
+- A `--shuffled` option to several scripts for reading from `*iloci.shuffled.tsv`, and improved access from the GenomeDB class.
 
 ## [0.3.7] - 2016-03-19
 ### Fixed